order 60%

+++ database/migrations/2015_07_09_042911_create_product_table.php
@@ -16,9 +16,9 @@ class CreateProductTable extends Migration
             $table->increments('id');
             $table->string('key_product')->unique();
             $table->string('name_product');
-            $table->integer('price_import');
-            $table->integer('price');
-            $table->integer('cost');
+            $table->bigInteger('price_import');
+            $table->bigInteger('price');
+            $table->bigInteger('cost');
             $table->string('image');
             $table->text('information');
